refactor: FSD Phase 6e — ノート編集を entities/note へ集約し共有 UI を配置 (FRESTYLE-171)
複数画面で共有される「ノート編集」の部品を entities/note へ集約し、純粋な
共有 UI を shared/ui へ移す。

- useNoteEditor → entities/note/model
  (entities/note の Note 型に依存。notes・course-detail・NoteMarkdownEditor
   の 3 者で共有。ノート編集の状態管理で note entity に密結合)
- NoteMarkdownEditor → entities/note/ui
  (entities/note に依存。notes・course-detail・markdown-syntax-help の
   3 画面で共有されるノートエディタ)
- PublicHeader → shared/ui
  (依存なしの純粋 UI。login・company-application の公開 2 画面で共有)
- useTeachingMaterialEditor → pages/course-detail/model
  (course-detail 専用。useNoteEditor(entities/note)を使う。pages→entities OK)

entities/note の Public API に useNoteEditor / SaveStatus / NoteMarkdownEditor
を追加。境界 lint が自己参照(useNoteEditor が Note 型を自 barrel から取得)を
検出したので相対に直した。

既存テストの assert は変更していない。テスト件数 1184 で不変。
