+++ frontend/src/pages/course-detail/model/__tests__/useTeachingMaterialEditor.test.ts
