+++ frontend/src/pages/course-detail/ui/CourseDetailPage.tsx
@@ -16,10 +16,10 @@ import EmptyState from '@/shared/ui/EmptyState';
 import FaviconIcon from '@/shared/ui/icons/FaviconIcon';
 import ConfirmModal from '@/shared/ui/ConfirmModal';
 import Loading from '@/shared/ui/Loading';
-import NoteMarkdownEditor from '@/components/NoteMarkdownEditor';
+import { NoteMarkdownEditor } from '@/entities/note';
 import MarkdownTableOfContents from '@/shared/ui/MarkdownTableOfContents';
 import { useTeachingMaterials } from '../model/useTeachingMaterials';
-import { useTeachingMaterialEditor } from '@/hooks/useTeachingMaterialEditor';
+import { useTeachingMaterialEditor } from '../model/useTeachingMaterialEditor';
 import { useChapterResume } from '../model/useChapterResume';
 import { useNextCourse } from '../model/useNextCourse';
 import { useLessonProgress } from '../model/useLessonProgress';
