+++ frontend/src/shared/ui/PublicHeader.tsx
